Add popup to say-hello UiAction
This change demonstrates the recently added JS RPC feature.

The `say-hello` action provides JS code hello.js, that is executed when
the user clicks on the "Say hello" action on the ChangeScreen2.

The old change screen still doesn't support JS RPC: the JS part is not
honored, only the UiAction apply method is executed.

The JS code shows a popup with a greeting message and language checkbox.
Clicking on the "Say hello" button in the popup sends a request to UiAction 
with the input parameters from the popup dialog.

Change-Id: Ie6abb40b00cfac9e6f86e9a37a8b2547d6ed7b50